+++ taiga/projects/wiki/validators.py
@@ -17,12 +17,15 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from taiga.base.api import validators
+from taiga.base.api import serializers
 from taiga.projects.notifications.validators import WatchersValidator
 
 from . import models
 
 
 class WikiPageValidator(WatchersValidator, validators.ModelValidator):
+    slug = serializers.CharField()
+
     class Meta:
         model = models.WikiPage
         read_only_fields = ('modified_date', 'created_date', 'owner')